requirements.txt
Flask
pandas
requests
smartsheet-python-sdk
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
openpyxl
waitress
